feat: Header 컴포넌트 정리 완료
