+++ src/router.tsx
@@ -2,18 +2,19 @@ import { createBrowserRouter } from "react-router-dom";
 import App from "./App";
 import Home from "./pages/Home";
 import Market from "./pages/Market";
-import Search from "./pages/Search";
+
 import Cart from "./pages/Cart";
 import Profile from "./pages/Profile";
-import CategoryDetail from "./pages/Search/categoryDetail";
+import CategoryDetail from "./pages/Category/categoryDetail";
 import MarketDetail from "./pages/Market/marketDetail";
 import MdRecommend from "./pages/Home/category/mdRecommend";
 import Discount from "./pages/Home/category/discount";
 import Login from "./pages/Login";
 import Signup from "./pages/Signup";
-import OnlySearch from "./pages/Search/onlySearch";
+import OnlySearch from "./pages/Category/onlySearch";
 import OnlyForCat from "./pages/Home/category/onlyForCat";
-import BestSeller from "./pages/Home/category/bestseller";
+import Search from "./pages/Category";
+import BestSeller from "./pages/Home/category/bestSeller";
 
 export const router = createBrowserRouter([
     {
